+++ src/components/layouts/layout.jsx
@@ -1,7 +1,7 @@
 // ヘッダーとフッターをLayoutコンポーネントで管理する
 
-import Header from './header/header'
-import Footer from './footer/footer'
+import Header from './header/Header'
+import Footer from './footer/Footer'
 
 export default function Layout({ children }) {
   return (
